requirements.txt
fastmcp>=0.4.1
requests>=2.31.0
python-dotenv>=1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1 